Fix code and tests compatibility with Cardinal update (#14)
This makes all tests pass again, in many cases the values have changed.

There were a few changes necessary to the code which were not included in my previous PR. I've tried to stay close to the original tests, but in a few cases I had to make some adjustments.

In general I understand the assertions are primarily the output of dput so I would simply recommend that we specify the tolerance value when performing assertions, to at least not be affected by simple floating point errors (e.g. on a different machine).